+++ code/game/g_local.h
@@ -362,10 +362,10 @@ typedef struct {
 	int			fragcarrier;
 	int			assists;
 
-	float		lasthurtcarrier;
-	float		lastreturnedflag;
-	float		flagsince;
-	float		lastfraggedcarrier;
+	int			lasthurtcarrier;
+	int			lastreturnedflag;
+	int			flagsince;
+	int			lastfraggedcarrier;
 } playerTeamState_t;
 
 // the auto following clients don't follow a specific client
@@ -770,7 +770,7 @@ Ghoul2 Insert End
 //
 qboolean CanDamage (gentity_t *targ, vec3_t origin);
 void G_Damage (gentity_t *targ, gentity_t *inflictor, gentity_t *attacker, const vec3_t directio, const vec3_t point, int damage, int dflags, meansOfDeath_t mod);
-qboolean G_RadiusDamage (vec3_t origin, gentity_t *attacker, float damage, float radius, gentity_t *ignore, meansOfDeath_t mod);
+qboolean G_RadiusDamage (vec3_t origin, gentity_t *attacker, int damage, int radius, gentity_t *ignore, meansOfDeath_t mod);
 void body_die( gentity_t *self, gentity_t *inflictor, gentity_t *attacker, int damage, meansOfDeath_t meansOfDeath );
 void TossClientWeapon(gentity_t *self, vec3_t direction, float speed);
 void TossClientItems( gentity_t *self );
